requirements.txt
appdirs==1.4.3
asgiref==3.7.2
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2019.11.28
chardet==3.0.4
cliapp==1.0.1
click==8.1.3
cmdtest==0.1.1
colorama==0.4.3
cryptography==2.8
dbus-python==1.2.16
defer==1.0.4
distlib==0.3.0
distro==1.4.0
distro-info==1.0
Django==4.2.2
django-bootstrap-v5==1.0.11
entrypoints==0.3
Faker==24.4.0
fasteners==0.14.1
filelock==3.0.12
Flask==2.3.2
future==0.18.2
httplib2==0.14.0
idna==2.8
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==1.3
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==2.1.2
monotonic==1.5
more-itertools==4.2.0
MouseInfo==0.1.3
mysql-connector-python==8.3.0
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
packaging==20.3
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
plyer==2.1.0
protobuf==3.6.1
PyAutoGUI==0.9.53
pycups==1.9.73
python-dateutil==2.7.3
python3-xlib==0.15
PythonTurtle==0.3.2
pytweening==1.0.4
pytz==2019.3
pywhatkit==5.4
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
scapy==2.5.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
soupsieve==2.4.1
sqlparse==0.4.4
typing-extensions==4.6.3
urllib3==1.25.8
virtualenv==20.0.17
wadllib==1.3.3
Werkzeug==2.3.3
wikipedia==1.4.0
zipp==1.0.0
